Consolidate the implementation of createImageBitmap(HTMLVideoElement)
In the current layout test for createImageBitmap from an HTMLVideoElement,
we draw the created ImageBitmap onto canvas, and test whether a pixel
is opaque or transparent. This is not enough. We should compare the
created ImageBitmap with the result of calling ctx.drawImage(video).

This CL re-write the test, and discovered that there are bugs in our
current implementation. We fix the bug and update the layout tests.

Review-Url: https://codereview.chromium.org/2274553002
Cr-Original-Commit-Position: refs/heads/master@{#416556}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8989d24501a4b7964a44f71a619fbed789ef9e6e