[Worker] Call prepareForShutdown ASAP when terminated forcibly
When a worker is terminated forcibly (i.e., TerminateExecution
is called), it is possible that many tasks are preceding
performShutdownOnWorkerThread on the worker task queue. As the v8 context
is in an invalid state, we need to ask worker objects to stop working
before executing all the queued tasks.

BUG=641215

Review-Url: https://codereview.chromium.org/2299883006
Cr-Original-Commit-Position: refs/heads/master@{#416598}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7a2a29de535c01c40ef8ee11621064572676e61d