webkitpy: Make all regex strings use raw string literals.
This fixes the anomalous-backslash-in-string pylint warnings.

The advantage of using raw string literals is that it makes the behavior
of backslashes a little easier to understand; with normal string literals,
"\a\b\c" is "<BEL><BS><backslash><lower-case-c>", whereas with raw
string literals, each of those backslashes is just a backslash character.

The advantage of this pylint warning in general is that enforces
escaping literal backslashes in regular strings, which makes it
clear that a backslash was intended rather than an escape sequence.

BUG=643937

Review-Url: https://codereview.chromium.org/2305143003
Cr-Original-Commit-Position: refs/heads/master@{#416636}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7beb14b269407802ce44aad29b778529ee473ed