Rename blink::GraphicsContext::colorFilter() to getColorFilter().
The rename is needed to avoid a naming collision after changing from
Blink to Chromium naming style.  Right now we have a |ColorFilter| type
and a |colorFilter| accessor method (differing by case of the first
character); after a naive rename by the rewrite_to_chrome_style tool we
would end up with |ColorFilter| being the name of both the type and the
accessor method (with both living in the same namespace).

Prepending a "get" prefix to the name of the accessor method is the
workaround that fits into the guidance on the recommended
post-Blink-to-Chromium-rename style suggested by esprehn@ in
https://crbug.com/582312#c17:
- Getters favor not using "Get", ex. FirstChild()
- Unless the type name conflicts, in which case you can either rename
  the type if it's easy and makes sense, or add "Get", ex. GetContext().

BUG=582312
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2308793002
Cr-Original-Commit-Position: refs/heads/master@{#416697}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e24b508116acb1ea006d0c6ab5774cc05391c4f6