Disallow reentrance of FrameView::updateLifecyclePhasesInternal()
Reentrance of FrameView::updateLifecyclePhasesInternal() is bad.
For example, when we are invalidating paint, we call a function which
looks like a non-side-affecting getter which unexpectedly initiates
another lifecycle update which may unexpectedly delete the object
that the caller is processing.

This CL checks reentrance of FrameView::updateLifecyclePhasesInternal(),
and crash with NOTREACHED() if DCHECK is enabled or early return
otherwise.

BUG=590856,621360

Review-Url: https://codereview.chromium.org/2310453002
Cr-Original-Commit-Position: refs/heads/master@{#416712}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 66f753c15a71e28594ec9502b6d1f2b9c9d14dba