Plumb reserve method of the BudgetAPI
This CL adds the reserve method to both the budget_api mojo service and
to the BudgetService web platform API. This also provides a skeleton
implementation of the mojo service in chrome/browser, but the full
connection is left for a future CL. There are new LayoutTests and an
update to the BudgetService Javascript mock to test the new reserve
functionality.

This does provide some TODO cleanup, adding a success parameter to mojo
operations that could fail due to database errors, adding mappings from
strings to mojo values, and updating the LayoutTests to be more dynamic.

BUG=617971

Review-Url: https://codereview.chromium.org/2309863002
Cr-Original-Commit-Position: refs/heads/master@{#416905}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: df4af8684c8cee4e5b68691c03f62db259c5c07c