[css-grid] Fix performance regression in grid layout
Some of the changes added in crrev.com/415676 made layout of grid containers
a bit slower. That was because it modified some methods in GridTrackSize.h
which is a hot spot. That class is basically a cache of some values and it's
critical to keep it really fast. The aforementioned CL was adding an if
clause to a getter method, and that was causing the slowdown.

BUG=643359

Review-Url: https://codereview.chromium.org/2318993002
Cr-Original-Commit-Position: refs/heads/master@{#417001}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8b54987d15a92a54796a357daf6e021f0ff9a2b6