SVGImageForContainer::imageForCurrentFrame() needs to account for zoom
SVGImageForContainer tracks a container zoom factor independently of
the actual container size.  This zoom factor needs to be accounted for
in imageForCurrentFrame() -- otherwise we end up producing an SkImage
with a different size than reported by SVGImageForContainer::size().

The easiest way to accomplish this is to use size() when calling
SVGImage::imageForCurrentFrameForContainer(), instead of the unzoomed
m_containerSize.

BUG=643623
R=fs@opera.com,davve@opera.com

Review-Url: https://codereview.chromium.org/2323853002
Cr-Original-Commit-Position: refs/heads/master@{#417504}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8e62373579cf6eddb6723e7e8e39306cde20c934