Refactor async_test in RTCPeerConnection-ice layout test
In https://codereview.chromium.org/2273363004/ qyearsley pointed out that this test wasn't running in the browser. This refactors the test to pass a function argument to async_test. I was hoping to solve the browser problem via this refactor, but it is still unsolved.

BUG=614963

Review-Url: https://codereview.chromium.org/2316573004
Cr-Original-Commit-Position: refs/heads/master@{#417659}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2c961c0712e93b1c0c7a559415f71e838390ff6