Remove more blink modules gypi code.
Improves the templates for make_names and process_in_files so they can be used
in a few more situations (not automatically inserting deps), clean up callers
that were manually calling these.

Add a new target bindings/modules:generated which compiles the sources
generated by the scripts in that file. This eliminates the hardcoding of these
source lists.

With this, the only remaining blink work should be to move the rest of the
source files into the modules sub-directories.

Review-Url: https://codereview.chromium.org/2328773002
Cr-Original-Commit-Position: refs/heads/master@{#417667}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0a91363934d878f3cd10aed2b4caf6fe239ec571