Temporarily stop honoring text-size-adjust as it breaks accessibility
This patch temporarily diables text-size-adjust from affecting the
text autosizer because the accessibility font size adjustment also
uses the text autosizer and sites should not be able to disable the
accessibility settings.

This is a temporary change. A proper fix will be coming in a followup.

Text-size-adjust support was added in:
https://crrev.com/5365b8ad85c2f5cee6034b8baea3c764ed78e1ea

BUG=645269

Review-Url: https://codereview.chromium.org/2330513002
Cr-Original-Commit-Position: refs/heads/master@{#417738}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4e8989e66c9cb0e427ea886dfc5ef51ef0e4e836