Move AX local bounding box calculation into LayoutObject
Minor refactoring. Now that we've successfully switched
to using relative bounding boxes in the accessibility tree,
it's cleaner to have AXObject call a single interface on
LayoutObject rather than calling different code depending
on the type of LayoutObject.

This will also help ensure this doesn't break the next
time the layout code is refactored.

This is covered by existing tests, in particular
accessibility/bounds-calc.html tests a few examples of each
of the four types of LayoutObject that need to implement
this method.

BUG=618120

Review-Url: https://codereview.chromium.org/2324603002
Cr-Original-Commit-Position: refs/heads/master@{#417775}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 64a061448438227cdd16d73d4d6944de6c5ddfd7