LayoutTable: gracefully degrade in the presence of collspan > 1.
Currently, there's a boolean field that allows to take a fast path
when mapping absolute columns to effective columns (and back), but
the fast path is completely abandoned, if a single call with
colspan > 1 presented.

This change allows gracefully degrade the performance of the fast
path by maintaining the number of first cells with colspan = 1,
which allows to skip them, even if some cells have colspan > 1.

This change speeds up large-table-with-collapsed-borders-and-colspans-wider-than-table
microbenchmark by approximately 2x, see:
https://storage.googleapis.com/cfi-stats/2016-09-09/results.html

BUG=643724

Review-Url: https://codereview.chromium.org/2325953002
Cr-Original-Commit-Position: refs/heads/master@{#417829}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 60f47383d88f942923b5ff963023d4547397acc2