Conditional hittesting and sending boundary events for captured pointer events
This CL adds the support sending boundary events for captured
pointer events if needed when pointerEventV1SpecCapturingEnabled
flag is enabled. Although it might not have captured all edge
cases but it does satisfy the basic needs and normal
use cases which suffices the testing need.
This CL does not change any behavior when the flag is off and
it is only for testing purposes when the flag is on.
Removal of this flag is tracked in crbug.com/642776.

BUG=640700

Review-Url: https://codereview.chromium.org/2326653003
Cr-Original-Commit-Position: refs/heads/master@{#418039}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ebbc77fb19301347fe7d062691ad20d82d5eb865