Deflake css3/filter/effect layout tests
https://codereview.chromium.org/1960543002 introduced some flakiness in
these tests, which appears to be due to texture sampling flakiness in
Mesa.  One workaround for this is to resize the bitmap in the tests to
be 256x128.  Because the tile (and the tile draw quad) will use this
size directly (instead of rounding up to some other size), the tile
and the render pass share the same size and are nice round numbers,
so texture sampling should have no errors.

BUG=621892

Review-Url: https://codereview.chromium.org/2304243002
Cr-Original-Commit-Position: refs/heads/master@{#418428}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3950059021388215d3864e552c7da761d4c3f822