Handle newCursorPosition correctly for Android's commitText()
Android’s InputConnection API specifies a way to commit (confirm) the
current composition and move the cursor at the same time:
https://developer.android.com/reference/android/view/inputmethod/
InputConnection.html#commitText(java.lang.CharSequence, int)

Currently, commitText doesn't conform to the documentation nor to
Android TextView's behavior when newCursorPosition is not 1.
newCursorPosition
is a relative value that depends on the location of the text being committed.

When |newCursorPosition| < 1, the cursor should move to the left of
the committing text by |newCursorPosition| characters. If the cursor
exceeds the left boundary, it will clamp to  the left boundary.

When |newCursorPosition| >= 1, the cursor will move to the right of
the committing text by |newCursorPosition - 1| characters. If the cursor
exceeds the right boundary, it will clamp to the right boundary.

(Rebased version of yabinh@'s http://crrev.com/1995333002 which received
lgt'ms from TBRed reviewers below, I'm landing this on his behalf.)

TBR=tkent,yosin,changwan,sievers,yabinh
NOTRY=true
BUG=570920

Review-Url: https://codereview.chromium.org/2339793002
Cr-Original-Commit-Position: refs/heads/master@{#418475}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 87b8f7c3fc3e5bbcc9601023df0fe79cb79234e8