Restore a collapsed leading space of text used for line break
When a text is wrapped during layout, the leading space of the
text can be collapsed and a line break is inserted instead of
the space. In this case, we need to restore the collapsed space
when we copy the text.

This CL handles the case that the below CL didn't cover:
https://codereview.chromium.org/2320533002/

In addition, while iterating through the DOM range, the below
case with the TextIteratorBehavior flag don't need to restore
the leading space.

* Running DumpRenderTree : TextIteratorForInnerText
* Searching text : TextIteratorDoesNotBreakAtReplacedElement
* Getting a plain text for copy & paste: TextIteratorEmitsImageAltText

BUG=318925
TEST=editing/pasteboard/restore-collapsed-space-for-copy.html

Review-Url: https://codereview.chromium.org/2325553002
Cr-Original-Commit-Position: refs/heads/master@{#418557}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 131df048e7e445a03706078367285112e6b3467d