Disable clipping on root scroller's ancestors.
Currently, while the user is hiding the top controls, the compositor expands the
main frame's clipping layer so that we reveal more content. This has to happen
because we don't actually resize the anything in the Layer/Layout trees in
Blink until after the top controls are fully shown or hidden.

When an element is set as the root scroller, we want to provide the same
effect. However, since there's still masking layers above the root scroller, we
can't see the expanded clip.  Instead of expanding all the layers above the
root scroller, this CL removes the masksToBounds bit on all layers above the
root scroller. This means parent boxes with hidden overflow will not clip the
root scroller but this is acceptable as a property of the root scroller.

Also, GraphicsLayer setMasksToBoundaries kept a bit on GraphicsLayer but this was
never read so I removed it.

BUG=505516
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2289833002
Cr-Original-Commit-Position: refs/heads/master@{#418647}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 469bcf04ced0868049864010a367b0d56a73ffa7