[layoutng] Set the intrinsic content logical height on the old layout tree
The overflow size is the content size without being affected by width/height
and max/min properties, so it's exactly what we need here.

This is important because without this change, there are DCHECK failures
if we're a child of a flexbox as the size remains at -1.

R=eae@chromium.org,ikilpatrick@chromium.org,glebl@chromium.org
BUG=635619

Review-Url: https://codereview.chromium.org/2337763005
Cr-Original-Commit-Position: refs/heads/master@{#418695}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 982432d01a43f9bede4b7b442f9b7d5ffa8a20c2