Enable the scroll of empty elements by setting the rect's size to 1.
LayoutBox::scrollRectToVisible will return if the new rect to be scrolled is
empty after clipped by layerBounds. However, the rect to be scrolled might be
an empty rect originally. A common case is empty <a> elements used as named
anchors for navigation. This patch sets the size of the empty rect to 1 so that
it can still be scrolled.

BUG=646738

Review-Url: https://codereview.chromium.org/2339653003
Cr-Original-Commit-Position: refs/heads/master@{#418856}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6d73ab5ed5d9b00bc48da255dc6599e90e72f599