Move TopDocumentRootScrollerController to a separate object on FrameHost
It contains a number of methods that don't conceptually belong on
RootScrollerController. This is also a natural step towards making this feature
work with out of process iframes.

The main changes in this CL are:

  -TopDocumentRootScrollerController is no longer a subclass of
   RootScrollerController

  -TopDocumentRootScrollerController is now owned by FrameHost

  -All the methods on RootScrollerController that really only belonged on
   the Page level object have been removed.

The class will be renamed in a followup CL to be more representative of these
changes.

BUG=505516
CQ_INCLUDE_TRYBOTS=master.tryserver.chromium.linux:linux_layout_tests_slimming_paint_v2

Review-Url: https://codereview.chromium.org/2285253003
Cr-Original-Commit-Position: refs/heads/master@{#418868}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 23cd9757ba798fa41b4d8dd80eb8457da6a9fa9e